requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.10.1
requests==2.20.0

